+++ cpp/src/arrow/pretty_print-test.cc
@@ -31,7 +31,7 @@
 #include "arrow/test-util.h"
 #include "arrow/type.h"
 #include "arrow/type_traits.h"
-#include "arrow/util/int128.h"
+#include "arrow/util/decimal.h"
 
 namespace arrow {
 
@@ -115,12 +115,12 @@ TEST_F(TestPrettyPrint, DecimalType) {
 
   DecimalBuilder builder(type);
 
-  Int128 val;
+  Decimal128 val;
 
-  ASSERT_OK(Int128::FromString("123.4567", &val));
+  ASSERT_OK(Decimal128::FromString("123.4567", &val));
   ASSERT_OK(builder.Append(val));
 
-  ASSERT_OK(Int128::FromString("456.7891", &val));
+  ASSERT_OK(Decimal128::FromString("456.7891", &val));
   ASSERT_OK(builder.Append(val));
   ASSERT_OK(builder.AppendNull());
